fix(turborepo): Remove automatic linking. (#5776)
The automatic `link` (which only occurs for `--sso-team`) does not
account for the fact that `login` is a one-time, global operation, which
can be run from _anywhere_ because of global `turbo`.

This means that, _wherever_ you run `turbo login --sso-team` you get a
`.turbo/config.json` generated relative to `cwd`.

The preferred method for automatically handling linking should be
storing the remote cache configuration inside of `turbo.json`—which
makes this automatic `link` step irrelevant and enables us to test for
cache connectivity during runs and know concretely whether or not the
exhibited behavior is intentional.

First step: remove the clearly buggy `link` implementation.

Review in whitespace-only mode:
https://github.com/vercel/turbo/pull/5776/files?w=1

Closes TURBO-1234