chore: require lockfile implementations to be threadsafe (#5781)
### Description

In order to safely share access to the `PackageGraph` between threads,
we need all the types it holds to be threadsafe. The berry lockfile was
the only part holding us back due to it's use of `Rc` to avoid cloning
some data (this data was borrowed before we removed it's self
referential nature).

Alternatives:
- Not requiring `Lockfile` implementers to be `Send + Sync` and instead
hold a `Box<dyn Lockfile + Send + Sync>`: requiring lockfiles to be
`Send + Sync` will allow us to parallelize the walking of the lockfile
in the future
- Replacing `Rc` usage with `Arc` usage in berry lockfile: For the
`BerryPackage` struct clone it exactly once from the `LockfileData`
struct, the cost of `Arc` seemed to outweigh the savings of avoiding
that clone. The `Rc` around the `LockfileData` was only used to avoid
cloning the entire lockfile struct when we pruned it via `turbo prune`.
Since this only happens in `turbo prune` where FS operations make up the
bulk of the runtime, I'm not concerned with this perf hit.

### Testing Instructions

Existing unit tests.
I used `cargo instruments --example berry_resolutions -t time` to see if
this was any large perf hit and it seems this like maybe ~1ms.


Closes TURBO-1239

Co-authored-by: Chris Olszewski <Chris Olszewski>