make swc comments immutable (#5784)
### Description

This fixes a consistency problem. SWC Emitter is mutating the passed
`Comments` while emitting (probably to avoid cloning).
This is bad for us as it mutates a value that should be immutable (as
it's in the cache). It causes missing comments when generating a module
again or with different configs.

This change changes it to force comments to stay immutable after being
parsed.

Closes WEB-1425