feat: port output control (#5906)
### Description

This is a port of what we do in Go for controlling writing to
stdout/stderr. This abstraction will be used in conjunction with
`PrefixedUI` and `LogWriter` to achieve the same behavior as Go.

In Go we achieve this behavior with a variety of wrapper structs around
stdout and stderr that implement `io.Write`:

![graph](https://github.com/vercel/turbo/assets/4131117/49247730-eeb1-4b95-8533-743f212a9127)

In Rust we combine all the the nodes from "concurrent ui" on down and
expose `OutputWriter` which implements `io::Write` which can then be
wrapped in either of the other structs this crate provides. This roughly
translates to [these lines in
Go](https://github.com/vercel/turbo/blob/main/cli/internal/run/real_run.go#L212-L235)

Notes:
- Currently this we wrap the underlying output writers in an
`Arc<Mutex>` to allow them to be written to from multiple threads. This
matches the behavior in Go, but in the future if we find we have too
much contention with the mutex, we could try switching to a mpsc
channel.
- This PR also includes adding a separate writer for errors to
`PrefixedUI` to match the Go implementation.

### Testing Instructions

Added unit tests for the new `output` module as well as an integration
test that displays how `turbo` will use all of these modules together to
achieve the same behavior as we have in Go. The integration test models
how we use these components in Go: [prefix
ui](https://github.com/vercel/turbo/blob/main/cli/internal/run/real_run.go#L412-L418)
and [task
writer](https://github.com/vercel/turbo/blob/main/cli/internal/run/real_run.go#L412-L418)


Closes TURBO-1312

---------

Co-authored-by: Chris Olszewski <Chris Olszewski>