fix(ci): remove build_turborepo dependency from JS and examples tests (#6066)
JS and examples tests do not use a local build of the turbo binary, so
we do not need to build it first. More importantly, becuase the
build_turborepo job runs conditionally only when Go or Rust files are
changed, it prevents JS and examples tests from running when only
examples or JS files have changed.