Ensure process.env.TURBOPACK is always set (#6116)
### Description

Related to https://github.com/vercel/next.js/pull/56496. Server Actions
currently shows an error as it seems to use the react-dom-webpack
version instead of react-dom-turbopack. This seems to be caused by the
changes in https://github.com/vercel/next.js/pull/56438 which checks for
`process.env.TURBOPACK` which is not always set currently, unlike
`process.turbopack`. This PR adds it in the other places where
`process.turbopack` is set.

<!--
  ✍️ Write a short summary of your work.
  If necessary, include relevant screenshots.
-->

### Testing Instructions

<!--
  Give a quick description of steps to test your changes.
-->