Revert "Reuse existing login tokens" (#6133)
This reverts #6034 and #6115 which built on top of it.

The attempt to read the `token_path` doesn't actually parse the contents
or otherwise attempt to grab the token from the configuration file.

There is an additional issue in that it doesn't account for existing
tokens in any of the other configuration sources. We already have a
full, parsed configuration at the time we call into the auth crate. It
should instead simply pass an `Option<&str>` of the existing configured
token into the function, not the `token_path`.

This also addresses @mehulkar's concerns about `token_path`.

Closes TURBO-1438