feat(turborepo): Layered Config (#5796)
This PR enables existing configuration settings to be stored in any of
"global", "local", or "shared" locations.
- Removes the config crate.
- Sets up configuration layering per our requirements.
- Does not make any behavior changes as to where link/unlink and
login/logout store the team and token values.

The primary value enabled here is that we can set `teamId` inside of
`turbo.json` and omit the `link` step, and make login aware of which
team scope it needs to auth to, eliminating the need for `--sso-team` in
most cases.

Closes TURBO-1252

---------

Co-authored-by: Nathan Hammond <Nathan Hammond>