chore(ci): Break up e2e tests (#6169)
This commit moves test definitions around and doesn't meaningfully
change them. The main change worth noting is that the `basicPrune`
and `explicitDepPrune` pipeline fixtures were running all 8 tests
with yarn. After this change they are only running the relevant prune
tests (both from a root directory and when the monorepo is nested
in a subdirectory). The extra tests are removed, because the changed
pipeline definition does not meaningfully test anything different in 
turbo run commands.