feat(benchmark): send TTFT benchmarks to slack
Breaks up benchmarking script so sending to TinyBird and Slack can
happen in parallel after all profiling is done. One downside of this is
that if the benchmark fails on any of the OS's in the matrix, no data
will be sent anywhere. This is good, because as we visualize this data
it will be important to have consistent data sets for each commit that
gets benchmarked.