Update Turbopack Docs (#6268)
This PR updates the Turbopack docs to reflect better where we are today:

* I have removed references to the sierpinski triangle benchmark. It was
somewhat useful as a contrived app early on in Turbopack's development,
but now that we support most features, we want our benchmarks to run
against real-world Next.js applications. We will publish benchmark
results running against [https://vercel.com/](https://vercel.com/) as we
get closer to becoming stable in Next.js
* I have removed the comparisons section for now. Turbopack is only
available in Next.js, so the only other bundler worth comparing right
now is webpack. Since the bundler is abstracted away in Next.js, the
main difference worth talking about will be performance, which we will
cover when we add the real-world benchmarks mentioned above.
* I edited the "Why Turbopack?" section to make it more clear why we
chose to build Turbopack for Next.js. The main reason that we neglected
to mention before is that existing bundlers were not built from the
ground up to support paradigms like SSR and RSC. We wanted to design
something from the ground up that we could influence to support modern
framework use cases well. We made some different choices than other
bundlers, but those choices are secondary and not worth focusing on at
this time.

Closes WEB-1844