fix: use which for package manager binary discovery (#6370)
### Description

I forgot that on Windows we need to use `which` to find binaries before
running them. We've run into this before and `which` is a straight
forward fix: https://github.com/vercel/turbo/pull/5208

### Testing Instructions

On Windows we now correctly find package manager binaries in the Rust
codepath. Verify it doesn't break non-Windows as well.


Closes TURBO-1586

Co-authored-by: Chris Olszewski <Chris Olszewski>