refactor(turborepo): Changed cache misses to be a non-error (#6369)
### Description

Originally we had cache misses be an error. That was a mistake because
it was too easy to propagate a cache miss up the stack and not handle it
properly. It was also too easy to not convert the error into a cache
miss. Now we have a `CacheResult` type that indicates either a miss or a
hit.

### Testing Instructions

<!--
  Give a quick description of steps to test your changes.
-->


Closes TURBO-1585

---------

Co-authored-by: nicholaslyang <Nicholas Yang>
Co-authored-by: Chris Olszewski <chris.olszewski@vercel.com>