chore(tests): replace fixture with a fixture config (#6391)
This commit deletes a very slightly different fixture monorepo in favor of a
fixture turbo.json config. In addition to reducing bloat, this will also make
it easier to add a Windows-customized config for the absolute path warning test,
since that requires `\\` separated paths to emit the warning on Windows.

This commit also establishes a directory location for turbo.json fixtures, and
moves some other custom configs into that directory.