chore(test): Remove single/double quotes in npm scripts in fixtures
When running these tests in Windows, these single and double quotes get printed
whereas on Darwin/Linux, they do not. This bug is repoted in npm
https://github.com/npm/cli/issues/6968. It would be possible to workaround
this by matching output in our prysk tests with regex and making the quotation
marks optional, but for most tests that is additional noise that we do not
need, so it is simpler to remove the quotes instead.