chore(ci): launch prysk tests with JS instead of bash (#6463)
We need to encode some Windows-specific logic when launching these tests
and it's easier to do it in JS than bash (although technically feasible
to do it in Bash). A JS script also lets us apply tooling like linting/formatting
that we don't have setup for bash.

The main drawback is that that this will launch prysk in a
subprocess, rather than sourcing a bash script in the same process.