feat(ci): run integration tests on windows (#6279)
This commit enables our integration tests on Windows and modifies
the tests to pass on on Windows. Mostly the issues are around finding
binaries with `.exe` extension (the test runner and turbo itself), file path
separators, and line endings.

This PR adds a new Github Action "job" for both the Go and Rust
variants, because there are still a few remaining tests that are not passing
that will need a little more investigation. Once this diff is merged, we can
individually look at each test, and make these tests blocking for PRs.