refactor: switch daemon version to be independent of turbo version (#6528)
This change is backwards compatible, since we set the starting version
of the daemon protocol to be equal to the current turbo version meaning
we will not get any clashes with old clients reporting a bad version.
Even though the version is currently equal to the turbo version, it is
not expected to always be the case.

### Description

If we want to allow other tools to interact with the turbo daemon, we
should not be as strict with our versioning. As it stands, a 3rd party
tool such as an LSP would need to perfectly match the version of turbo
used in the repo which is an unreasonable constraint.

This also relaxes the version requirement for all _new_ versions to
least the same minor version (ie. that all the features available to the
client are implemented by the server). Old clients will still demand
identical versions (and have that demand respected correctly).

### Testing Instructions

Unit tests for the semver check included.