chore(tests): fix daemon verbosity test on Windows (#6582)
This test fails on Windows because the log lines are a little different.
This may indicate that the behavior is different between platforms (for
example, there are 3 lines "No local turbo binary found" instead of 2),
but this test does not exist to assert against these differences. This
test is only to make sure that the env vars and flags for verbose
logging are respected. I have modified the test to remain true to its
intention and pass on Windows also.