chore(tests): fix absolute path warnings test on windows (#6564)
There are three bugs causing this test to fail:

- The regex in the test itself wasn't correct. This should be fixed after this commit.

- When `inputs` contains an absolute path, it fails glob parsing (both
Rust and Go codepaths). This is fixed by removing the `inputs` config
entirely. We should fix that separately, but don't need to block this
test on that, since it's for an unsupported feature.

- The Rust codepath fails to parse a Windows style absolute path
in globalDependencies entirely. This does not reproduce in the Go codepath,
so we should fix it, but it is out of scope here.