fix(turborepo): make linked installs test pass on Windows (#6477)
There are two fixes in this commit:

1. On Windows, canonicalizing a path that traverses a directory
and includes symlink behaves differently. So this commit breaks
up both steps into the order we need to find the turbo binary.

2. It updates the test case setup so that a symlink is created
successfully in Windows. The existing "cp -a" operation was following
the symlink and copying the target instead of copying the symlink
itself. rsync or other flags on cp did not change this behavior so
we had to recreate the symlink manually in the test setup. Note that
there are other symlinks in this fixture (which replicates a linked install
from pnpm and npm), but we do not recreate those symlinks because
they do not affect the test.