turbopack-css: fix rule duplication
This fixes an issue in css chunk generation where rules could be duplicated, which was likely introduced in the chunking refactor (a series of PRs starting with #6104).

CSS chunk generation had special logic that would walk the module graph on its own and include code. Combined with the topological list of modules provided by chunking, this unnecessarily included duplicate rules in incorrect order.

Test Plan: Added and updated snapshot tests.