feat(telemetry): generic parent to child event (#6830)
### Description

Previously, the parent<>child relationship between telemetry events
required that both events be the same event type. This was overly
restrictive, and this PR now allows any event type to be linked
together.

This is accomplished by requiring that the event passed to `with_parent`
implements the `Identifiable` trait, instead of it being the exact type
of a generic.

This also adds a few other events in and sets up piping for others.