Remove old redirects. (#6871)
### Description

As we work through some restructuring of the documentation, there will
be some rewrites that need to happen.

Here, I'm cleaning things up so our redirects stay as simple and easy to
read as possible. I checked Search Console and none of the source paths
in the list are currently being indexed by Google.