fix aggregation of outdated children and collectibles (#6885)
### Description

The delayed removal of children and collectibles for "in progress" tasks
of the aggregate tree causes some inconsistency in the aggregated tree
when "in progress" task are connected to other node. We need to fix that
by reporting outdated children and collectibles as children/collectibles
of the "in progress" tasks.

This fixes an race condition when temporary errors are incorrectly shows
as final result. e. g. it was visible as `can't resolve
"ACTIONS_MODULE9"` in next.js.


Closes PACK-2192