revert auth json (#6883)
We're currently blocked from shipping due to issues found with the
`auth.json` code. We believe all of the issues have been addressed, but
we should do some due diligence and dogfood a canary for a bit before
cutting a release.

We have some fixes that we need to get out and want the release to be
solid without accidentally introducing some auth related regressions.

I believe these are all the auth.json related PRs and gets us back to a
known good state.

Closes TURBO-1989