Improve docs search index. (#6878)
### Description

Previously, our docs search had a mixed bag of problems that made
discoverability poor. It was difficult to:
- see what content we were really missing
- sift through results that didn't make sense for the query
- find results that should be ranked more highly
- filter out results that don't contain useful information

This PR improves our search by leveraging
[Pagefind](https://pagefind.app) to iterate on the default search
provided by Nextra. The new search input and dropdown is meant to be a
visually faithful recreation of that dropdown with a few small liberties
taken on design.


### Testing

Visit the preview and try as many useful queries as you can think of.
Non-exhaustive list of things to look for:
- Are hashlinks working as expected?
- Does the search input focus and blur when you'd expect? Are the search
results removed from render when you'd expect?
- Is it fast?
- Are results missing that you would expect to see?
- Any visual bugs with responsiveness?
- Any visual bugs with dark/light mode?

CLOSES TURBO-1862